fix getTotalClassSession() and Add viewStudentProfile() Api

+++ src/main/java/com/atcollabo/hackathon/classdojo/service/impl/TeacherServiceImpl.java
@@ -34,7 +34,7 @@ public class TeacherServiceImpl implements TeacherService {
 
     @Override
     public int getTotalClassSessions(Long classId) {
-        return teacherRepository.getTotalClassSessions(classId);
+        return (int)teacherRepository.getTotalClassSessions(classId);
     }
 
     @Override
